Refactor find_cconj_heads
